Add stateful capability to the builder
This reverts the changes in 3db02e back to a straight `include` of each
inclusion module, instead of deferring the `include` to the
`Module#included` hook. This seemed more straightforward than an
alternative route: allowing only `#call`able defined hooks.

`#call`able defined hooks would be arguably more powerful than the
current implementation of defined hooks, but would be cognitively more
difficult to work with because they would have to modify the module from
a non-owning perspective. This would generate code that is harder to
write, harder to read, and harder to debug when things go wrong.

I might reevaluate this at a later date, but this seemed like the best
option for the moment.