Add a Buildable module for mixing into modules
By mixing in `Buildable`, you gain the ability to send `::new` to a
module with a configuration state to build a new version of the module.
This enables you to quickly reuse a `Builder` through a succinct
interface without worrying about explicitly creating the builder object
and using its built module.