Remove the unnecessary build context class
At this time, there was no extra information being passed to the build
context; it was merely a delegator to the wrapped builder. That is
unnecessary complexity, as revealed by mutation testing. Let's remove it
for now and just yield the builder, because chances are the build
context was a YAGNI violation.